back/requirements.txt
asgiref==3.2.9
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-anymail==7.1.0
django-cors-headers==3.4.0
django-environ==0.4.5
djangorestframework==3.11.0
idna==2.9
Pillow==7.1.2
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
